# shared/contracts/openapi/batch-downloader-api.yaml
openapi: 3.1.0
info:
  title: Batch Downloader API
  version: 1.0.0
  description: API pública versionada para identidad, catálogo y trabajos de descarga.
  license:
    name: GPL-3.0-only
    identifier: GPL-3.0-only
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /admin/auth/login:
    servers:
      - url: /
    post:
      operationId: adminLogin
      summary: Iniciar sesión administrativa por username
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: Sesión ADMIN iniciada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/auth/me:
    servers:
      - url: /
    get:
      operationId: currentAdmin
      summary: Consultar la sesión ADMIN
      responses:
        '200':
          description: Administrador autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/auth/logout:
    servers:
      - url: /
    post:
      operationId: adminLogout
      summary: Cerrar la sesión administrativa
      responses:
        '204':
          description: Sesión invalidada.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: BATCH_SESSION
    internalServiceToken:
      type: apiKey
      in: header
      name: X-Internal-Service-Token
...
  responses:
...
    Unauthorized:
      description: Sesión ausente o inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: La sesión o el estado de la cuenta no permiten la operación.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    AdminLoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
...
    User:
      type: object
      required: [id, username, email, emailVerified, role, notifyOnJobCompletion, createdAt, authenticationMethods]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        role:
          type: string
          enum: [USER, ADMIN]
        notifyOnJobCompletion:
          type: boolean
        createdAt:
          type: string
          format: date-time
        authenticationMethods:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [LOCAL, GOOGLE]
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true